feedback/requirements.txt
cassandra-driver
Flask==2.2.2
